requirements.txt
absl-py==0.9.0
astor==0.8.1
boto==2.49.0
boto3==1.12.46
botocore==1.15.46
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
docutils==0.15.2
Flask==1.1.2
gast==0.3.3
gensim==3.8.2
grpcio==1.28.1
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.9.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
mock==4.0.2
numpy==1.18.3
protobuf==3.11.3
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==1.11.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
